Move MusicBrainzAPI to a shared util
